backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.13.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
groq==0.4.1
httpx==0.25.2
pytz==2023.3
sendgrid==6.11.0
apscheduler==3.10.4
jinja2==3.1.2
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.5.0
sentence-transformers>=2.2.2
torch>=2.0.0